Ransom Note
This problem can be solved by using a combination of sets, dicts, and or maps. By using one of these data structures it just requires a traversal of the ransom note and checking it there is enough corresponding characters in the magazine, or if it exists in the magazine in general.