Find the Town Judge
This problem is solved by making two sets, one with all the people who trust another and one with all the people that are trusted by others. After this, subtracting the two sets and checking if the only element  (-1 if there are multiple elements) has a count of N - 1 is the answer.