Single Element in a Sorted Array
This problem is solved using a binary search to see when the pattern of the numbers is changed. In this problem the pattern is that at an odd index the pair will be one on the left and if the index is right then the pair will be one to the right. When there is only element, the pattern is messed up, thus having to find when the pattern is messed up.